Deleting by iterating the students.

+++ app/students/students.js
@@ -19,13 +19,18 @@ function($scope, $window, studentService) {
     // we could make this simpler by:
     // * changing the signature method for removeStudent(studentId)
     // * not displaying the name of the student we want to delete
-    studentService.getStudent(studentId).then(function(student) {
-      if ($window.confirm("Do you really want to delete '" + student["name"] + "'?")) {
-        studentService.removeStudent(student).then(function() {
-          studentService.getStudents($scope.search).then(function(students) {
-            $scope.students = students;
+    
+    // or we could also make this simpler by:
+    // * iterating the students to find one with matching id
+    angular.forEach($scope.students, function(student) {
+      if (student["id"] === studentId) {
+        if ($window.confirm("Do you really want to delete '" + student["name"] + "'?")) {
+          studentService.removeStudent(student).then(function() {
+            studentService.getStudents($scope.search).then(function(students) {
+              $scope.students = students;
+            });
           });
-        });
+        }
       }
     });
   }
